red flag list done
