+++ src/components/nav_bar/nav_bar.tsx
@@ -73,7 +73,7 @@ const NavBar = () => {
           <Link href={BFAURL.COMING_SOON}>{t('NAV_BAR.AUDITING_TOOL')}</Link>
         </div>
         <div className="text-white hover:text-primaryBlue">
-          <Link href={BFAURL.COMING_SOON}>{t('NAV_BAR.RED_FLAG')}</Link>
+          <Link href={BFAURL.RED_FLAG}>{t('NAV_BAR.RED_FLAG')}</Link>
         </div>
         <div className="text-white hover:text-primaryBlue">
           <Link href={BFAURL.COMING_SOON}>{t('NAV_BAR.FAQ')}</Link>
