+++ src/pages/app/chains/[chainId]/address/[addressId]/red-flag.tsx
@@ -1,25 +1,18 @@
 import Head from 'next/head';
 import Image from 'next/image';
-import {useState, useEffect} from 'react';
-import useStateRef from 'react-usestateref';
 import {useRouter} from 'next/router';
 import {useTranslation} from 'next-i18next';
 import {serverSideTranslations} from 'next-i18next/serverSideTranslations';
 import {GetStaticPaths, GetStaticProps} from 'next';
 import NavBar from '../../../../../../components/nav_bar/nav_bar';
-import SearchBar from '../../../../../../components/search_bar/search_bar';
-import SortingMenu from '../../../../../../components/sorting_menu/sorting_menu';
-import DatePicker from '../../../../../../components/date_picker/date_picker';
 import BoltButton from '../../../../../../components/bolt_button/bolt_button';
 import Footer from '../../../../../../components/footer/footer';
 import {BsArrowLeftShort} from 'react-icons/bs';
 import {getChainIcon} from '../../../../../../lib/common';
 import {TranslateFunction} from '../../../../../../interfaces/locale';
 import {dummyAddressData} from '../../../../../../interfaces/address';
-import {ITEM_PER_PAGE, sortOldAndNewOptions} from '../../../../../../constants/config';
-import Pagination from '../../../../../../components/pagination/pagination';
 import {IRedFlag} from '../../../../../../interfaces/red_flag';
-import RedFlagItem from '../../../../../../components/red_flag_item/red_flag_item';
+import RedFlagList from '../../../../../../components/red_flag_list/red_flag_list';
 
 interface IRedFlagOfAddressPageProps {
   chainId: string;
@@ -37,68 +30,6 @@ const RedFlagOfAddressPage = ({chainId, addressId, redFlagData}: IRedFlagOfAddre
   const router = useRouter();
   const backClickHandler = () => router.back();
 
-  // Info: (20231109 - Julian) Type Options
-  const flaggingType = redFlagData.map(redFlagData => redFlagData.redFlagType);
-  const typeOptions = ['SORTING.ALL', ...flaggingType];
-
-  // Info: (20231109 - Julian) States
-  const [search, setSearch, searchRef] = useStateRef('');
-  const [sorting, setSorting] = useState<string>(sortOldAndNewOptions[0]);
-  const [period, setPeriod] = useState({
-    startTimeStamp: 0,
-    endTimeStamp: 0,
-  });
-  const [filteredType, setFilteredType] = useState<string>(typeOptions[0]);
-  const [filteredRedFlagList, setFilteredRedFlagList] = useState<IRedFlag[]>(redFlagData);
-  const [activePage, setActivePage] = useState<number>(1);
-  const [totalPages, setTotalPages] = useState<number>(
-    Math.ceil(redFlagData.length / ITEM_PER_PAGE)
-  );
-
-  const endIdx = activePage * ITEM_PER_PAGE;
-  const startIdx = endIdx - ITEM_PER_PAGE;
-
-  const displayRedFlagList = filteredRedFlagList
-    // Info: (20231109 - Julian) pagination
-    .slice(startIdx, endIdx)
-    .map((redFlagData, index) => <RedFlagItem key={index} redFlagData={redFlagData} />);
-
-  useEffect(() => {
-    const searchResult = redFlagData
-      // Info: (20231109 - Julian) filter by search term
-      .filter((redFlagData: IRedFlag) => {
-        const searchTerm = searchRef.current.toLowerCase();
-        const type = redFlagData.redFlagType.toLowerCase();
-        const id = redFlagData.addressId.toLowerCase();
-        return searchTerm !== '' ? type.includes(searchTerm) || id.includes(searchTerm) : true;
-      })
-      // Info: (20231109 - Julian) filter by date range
-      .filter((redFlagData: IRedFlag) => {
-        const createdTimestamp = redFlagData.flaggingTimestamp;
-        const start = period.startTimeStamp;
-        const end = period.endTimeStamp;
-        // Info: (20231109 - Julian) if start and end are 0, it means that there is no period filter
-        const isCreatedTimestampInRange =
-          start === 0 && end === 0 ? true : createdTimestamp >= start && createdTimestamp <= end;
-        return isCreatedTimestampInRange;
-      })
-      // Info: (20231109 - Julian) filter by type
-      .filter((redFlagData: IRedFlag) => {
-        const type = redFlagData.redFlagType.toLowerCase();
-        return filteredType !== typeOptions[0] ? filteredType.toLowerCase().includes(type) : true;
-      })
-      // Info: (20231109 - Julian) sort by Newest or Oldest
-      .sort((a: IRedFlag, b: IRedFlag) => {
-        return sorting === sortOldAndNewOptions[0]
-          ? a.flaggingTimestamp - b.flaggingTimestamp
-          : b.flaggingTimestamp - a.flaggingTimestamp;
-      });
-
-    setFilteredRedFlagList(searchResult);
-    setActivePage(1);
-    setTotalPages(Math.ceil(searchResult.length / ITEM_PER_PAGE));
-  }, [filteredType, search, period, sorting]);
-
   return (
     <>
       <Head>
@@ -111,7 +42,7 @@ const RedFlagOfAddressPage = ({chainId, addressId, redFlagData}: IRedFlagOfAddre
         <div className="flex min-h-screen flex-col items-center overflow-hidden font-inter">
           <div className="flex w-full flex-1 flex-col items-center space-y-10 px-5 pb-10 pt-32 lg:px-40 lg:pt-40">
             {/* Info: (20231109 - Julian) Header */}
-            <div className="flex w-full items-center justify-start">
+            <div className="flex w-full items-start justify-start">
               {/* Info: (20231109 -Julian) Back Arrow Button */}
               <button onClick={backClickHandler} className="hidden lg:block">
                 <BsArrowLeftShort className="text-48px" />
@@ -133,48 +64,10 @@ const RedFlagOfAddressPage = ({chainId, addressId, redFlagData}: IRedFlagOfAddre
               </div>
             </div>
 
-            {/* Info: (20231109 - Julian) Search Filter */}
-            <div className="flex w-full flex-col items-end space-y-10">
-              {/* Info: (20231109 - Julian) Search Bar */}
-              <div className="mx-auto w-full lg:w-7/10">
-                <SearchBar
-                  searchBarPlaceholder={t('RED_FLAG_DETAIL_PAGE.SEARCH_PLACEHOLDER')}
-                  setSearch={setSearch}
-                />
-              </div>
-              <div className="flex w-full flex-col items-center gap-2 lg:flex-row lg:justify-between">
-                {/* Info: (20231109 - Julian) Type Select Menu */}
-                <div className="relative flex w-full items-center space-y-2 text-base lg:w-fit">
-                  <SortingMenu
-                    sortingOptions={typeOptions}
-                    sorting={filteredType}
-                    setSorting={setFilteredType}
-                  />
-                </div>
-                {/* Info: (20231109 - Julian) Date Picker */}
-                <div className="flex w-full items-center text-sm lg:w-fit lg:space-x-2">
-                  <p className="hidden text-lilac lg:block">{t('DATE_PICKER.DATE')} :</p>
-                  <DatePicker setFilteredPeriod={setPeriod} />
-                </div>
-                {/* Info: (20231109 - Julian) Sorting Menu */}
-                <div className="relative flex w-full items-center text-sm lg:w-fit lg:space-x-2">
-                  <p className="hidden text-lilac lg:block">{t('SORTING.SORT_BY')} :</p>
-                  <SortingMenu
-                    sortingOptions={sortOldAndNewOptions}
-                    sorting={sorting}
-                    setSorting={setSorting}
-                  />
-                </div>
-              </div>
-            </div>
-
             {/* Info: (20231109 - Julian) Red Flag List */}
-            <div className="flex w-full flex-col lg:mt-20">{displayRedFlagList}</div>
-            <Pagination
-              activePage={activePage}
-              setActivePage={setActivePage}
-              totalPages={totalPages}
-            />
+            <div className="w-full">
+              <RedFlagList redFlagData={redFlagData} />
+            </div>
 
             {/* Info: (20231109 - Julian) Back button */}
             <div className="">
